Improve `Remove` and `Add` commands
This commit improves the `Remove` command to allow it to take in an array of `alias`es to remove instead of just one. It can also still take in just one. Furthermore, this commit allows the user to pass in a `--description` to the `ally add` command, which is a first-try impl of adding documentation (as listed in the README).
# Full Changelog
- Add a `Dot` command with an `Open` subcommand to open the `.ally` file
    - Should we allow the user to have a base `.ally.config` file to configure `.ally` location?
- Still working on the `Ally.Dot.Format` command, which should format the `.ally` file.
- Add a `description` parameter to `Add`
- Allow the `Remove` command to take in an `Array` of aliases to remove
- Allow the `Remove` command to search `.zshrc` as well